chore: split files into services and routes
