+++ ingress/package.json
@@ -25,6 +25,7 @@
     "@types/uuid": "^9.0.8",
     "dotenv": "^16.4.5",
     "express": "^4.18.3",
+    "mongodb": "^6.5.0",
     "pg": "^8.11.3",
     "typescript": "^5.3.3",
     "uuid": "^9.0.1"
